V1.0 Company plea screens
Updates since last version 0.9.4

1.1.0 Start landing screen

Added additional bullet point
you must be the person, or a legal representative of the company, named
on the charge sheet.

1.2.0 Case details

New question that requires a yes or no answer
Are you making a plea on behalf of a company?

Yes - directs the user to the Your details for Companies
No - directs the user down the Your details screen

1.3.1 Your details for Companies
Company name
Company address
Position held within company – The user must be one of the following to
submit a plea on a companies behalf.
•	A director
•	The company director
•	The company solicitor
Contact number
Email address

When the user selects the continue CTA from the Your details for
companies, the user is directed to the plea screen. If the user pleads
guilty to a charge they are directed to the Financial details for
companies. Otherwise the user is directed to the review details screen

1.5.0 Company financial details
This screen is only shown if the user pleaded guilty to a charge.
Otherwise the user is directed to the review details screen
•	Number of employees
•	Gross turnover
•	Net turnover

Providing all required input fields have been completed correctly, when
the user selects the continue call to action they are directed to the
review details screen. Otherwise the system will display the relevant
error and instruction notification.

1.7.1a Review details Company guilty plea screen
Required data to be displayed in addition to existing data:
Case details
Require correct fields for company question which will be a yes or no
answer.

Your details
Company name
Company address
Position held within company – The user must be one of the following to
submit a plea on a companies behalf.
•	A director
•	The company director
•	The company solicitor
Contact number
Email address

Company financial details
Only shown if the user pleaded guilty
•	Number of employees
•	Gross turnover
•	Net turnover

Important - should read:
I confirm that I have read and understand the charges and the terms and
conditions. I also confirm am the person, or a legal representative of
the company, named in the Postal Requisition. The facts stated here are
correct to the best of my knowledge.

1.7.1b Review details Company not guilty plea screen
Required data to be displayed in addition to existing data:

Case details
Require correct fields for company question which will be a yes or no
answer.

Your details
Company name
Company address
Position held within company – The user must be one of the following to
submit a plea on a companies behalf.
•	A director
•	The company director
•	The company solicitor
Contact number
Email address

Important - should read:
I confirm that I have read and understand the charges and the terms and
conditions. I also confirm am the person, or a legal representative of
the company, named in the Postal Requisition. The facts stated here are
correct to the best of my knowledge.

1.11.1a Court email - company guilty plea
Required data to be displayed in addition to existing data:

Case details
Require correct fields for company question which will be a yes or no
answer.

Your details
Company name
Company address
Position held within company – The user must be one of the following to
submit a plea on a companies behalf.
A director
The company director
The company solicitor
Contact number
Email address

Company financial details
Only shown if the user pleaded guilty
Number of employees
Gross turnover
Net turnover

Important - should read:
I confirm that I have read and understand the charges and the terms and
conditions. I also confirm am the person, or a legal representative of
the company, named in the Postal Requisition. The facts stated here are
correct to the best of my knowledge.

1.11.1b Court email - company not guilty plea
Required data to be displayed in addition to existing data:

Case details
Require correct fields for company question which will be a yes or no
answer.

Your details
Company name
Company address
Position held within company – The user must be one of the following to
submit a plea on a companies behalf.
A director
The company director
The company solicitor
Contact number
Email address