Updates since version 1.0
Created 7 new screens for None JavaScript scenario

NJS 9.4.0 Your plea
NJS 9.4.1 Your plea continued
NJS 9.5.0 Your finances
NJS 9.5.1 Your finances continued - Employed
NJS 9.5.2 Your finances continued - Self-employed
NJS 9.5.3 Your finances continued - Receiving benefits
NJS 9.5.4 Your finances continued - Other

NJS 4.0 Your Plea
None JavaScript version of the Your plea screen

When the user selects either a guilty and not guilty option for each of
the pleas the screen does not update like the JavaScript version.
Instead the user selects the continue CTA which directs them to a new
screen NJS 4.1 Your Plea continued - where the details relating to the
previously selected options are shown.

Note
A link will be available to the user from NJS 4.1 Your Plea continued
screen that will enable the user to come back to this screen if they
want to change their plea.
If the user does go back and change their plea when they select the
continue CTA again the details displayed on the next screen NJS 4.1
Your Plea continued - should reflect that change.

NJS 4.1 Your Plea continue
None JavaScript version of the Your plea screen

Previously on NJS 4.0 Your Plea screen the user selected either a
guilty and not guilty option for each of the charges against them.
The details relating to the previously selected options are shown on
this screen

The screen displays the read only view of the previously selected plea
for that charge
A text link that when selected enables the user to go back from NJS 4.1
Your Plea continued screen back to NJS 4.0 Your Plea screen so they can
change their plea if required.
Note:
If the user does go back and change their plea when they come back to
this screen the details displayed should reflect that change.

Display the following information and input fields based on the
following selected option on NJS 4.0 Your Plea screen

Guilty
Special circumstances
What would you like the court to consider?
Text box

Not Guilty
Not guilty because?
Tell us why you believe you are not guilty.
Text box

NJS 9.5.0 Your finances
None JavaScript version of the Your finances screen

When the user selects an Are you? option the screen does not update
like the JavaScript version.
Instead the user selects the continue CTA which directs them to a new
screen

If the user selected Employed they will be directed to:
NJS 9.5.1 Your finances continued - Employed - where the details
relating to this option are shown.

If the user selected Employed they will be directed to:
NJS 9.5.2 Your finances continued - Self-employed - where the details
relating to this option are shown.

If the user selected Employed they will be directed to:
NJS 9.5.3 Your finances continued - Receiving benefits - where the
details relating to this option are shown.

If the user selected Employed they will be directed to:
NJS 9.5.4 Your finances continued - Other - where the details relating
to this option are shown.

Note
A link will be available to the user from either of these screens that
will enable the user to come back to this screen if they want to change
their selection.
If the user does go come back and change their selection, when they
select the continue CTA again the user should be redirected accordingly
based on their new selection.

NJS 9.5.1 Your finances continued - Employed
None JavaScript version of the Your finances screen after Are you
option was chosen and the user selected the Continue CTA.

Previously on NJS 9.5.0 Your finances screen the user selected the
Employed option
The details relating to this option are shown on this screen.

The screen displays the read only view of the previously selected option
A text link that when selected enables the user to go back from NJS
9.5.1 Your finances continued - Employed to NJS 9.5.0 Your finances
screen so they can change their previous selection if required.

Note:
If the user does go back and change their option, when they come back
to this screen the details displayed should reflect that change.

NJS 9.5.2 Your finances continued - Self employed
None JavaScript version of the Your finances screen after Are you
option was chosen and the user selected the Continue CTA.

Previously on NJS 9.5.0 Your finances screen the user selected the
Employed option
The details relating to this option are shown on this screen.

The screen displays the read only view of the previously selected option
A text link that when selected enables the user to go back from NJS
9.5.2 Your finances continued - self-employed to NJS 9.5.0 Your
finances screen so they can change their previous selection if required.

Note:
If the user does go back and change their option, when they come back
to this screen the details displayed should reflect that change.

NJS 9.5.3 Your finances continued - Receiving benefits
None JavaScript version of the Your finances screen after Are you
option was chosen and the user selected the Continue CTA.

Previously on NJS 9.5.0 Your finances screen the user selected the
Employed option
The details relating to this option are shown on this screen.

The screen displays the read only view of the previously selected option
A text link that when selected enables the user to go back from NJS
9.5.3 Your finances continued - Receiving benefits to NJS 9.5.0 Your
finances screen so they can change their previous selection if required.

Note:
If the user does go back and change their option, when they come back
to this screen the details displayed should reflect that change.

NJS 9.5.4 Your finances continued - Other
None JavaScript version of the Your finances screen after Are you
option was chosen and the user selected the Continue CTA.

Previously on NJS 9.5.0 Your finances screen the user selected the
Employed option
The details relating to this option are shown on this screen.

The screen displays the read only view of the previously selected option
A text link that when selected enables the user to go back from NJS
9.5.4 Your finances continued - Other to NJS 9.5.0 Your finances screen
so they can change their previous selection if required.

Note:
If the user does go back and change their option, when they come back
to this screen the details displayed should reflect that change.