BETA WP 1.8 Start screen update
0.1	Start landing screen
Updates since last version WP1.7

Change:
Make your plea online
Change to:
Make a plea online

Change:
Use this service to:
Plead guilty or not guilty to the driving offence(s) you're charged
with.

Change to:
Use this service to:
Plead guilty or not guilty for each traffic offence listed on the
charge sheet in the requisition pack.

Change:
By making your plea(s) online:
There's no need to come to court on your hearing date.

Change to:
After pleading online:
Do not come to court on the hearing date.

Change:
Before you start
you'll need the 'Postal Requisition' pack we sent to you
you can make your plea at any time up to the day before the court makes
a decision
you must be the person, or a legal representative of the company, named
on the charge sheet
entering your plea takes about 10 minutes
it is against the law to enter a plea for anyone else

Change to:

Before you start
you'll need information from the police requisition pack
entering a plea takes about 10 minutes
you must be the person, or an official representative of the company,
named on the requisition